finally delete volunteer with different projects

+++ pitch-for-good-api/src/main/java/org/wecancodeit/pitchforgood/controllers/VolunteerController.java
@@ -194,6 +194,7 @@ public class VolunteerController {
 	@DeleteMapping("delete/{volunteerId}")
 	public Collection<Volunteer> deleteVolunteer(@PathVariable Long volunteerId) {
 		Volunteer volunteerToDelete = volunteerRepo.findById(volunteerId).get();
+		System.out.println(volunteerToDelete);
 		
 		volunteerToDelete.removeSkillsInCollection();
 		volunteerToDelete.removeCausesInCollection();
@@ -202,27 +203,32 @@ public class VolunteerController {
 		
 		//get projects by that volunteer
 		Collection<Project> projects = projectRepo.findProjectsByVolunteer(volunteerToDelete);
+		System.out.println(projects);
 		
 		ArrayList<Organization> organizations = new ArrayList<>();
 		
-		for (Project project : projects) {
-			Organization organization = project.getOrganization();
-			System.out.println(organization.getProjects());
-			organizations.add(organization);
-
+		if (projects != null) {
 			
-			for (Organization org : organizations) {
-				org.removeProject(project);
-				projectRepo.delete(project);
+			for (Project project : projects) {
+				Organization organization = project.getOrganization();
+				System.out.println(organization);
+				System.out.println(organization.getProjects());
+				organizations.add(organization);
+				
+				
+				for (Organization org : organizations) {
+					if (org.getProjects().contains(project)) {
+						org.removeProject(project);
+						projectRepo.delete(project);
+					}
+				}
+				System.out.println(organization.getProjects());
 			}
-			
 		
-		System.out.println(organization.getProjects());
+		}
 		
 		volunteerRepo.delete(volunteerToDelete);
 		System.out.println(volunteerToDelete);
-		}
-
